Update showcart.php
For update quantity and delete.
Close for quantity. I am able to change the url, and get the information from url, need to be able to use the information users input not the count which is what it looks like he does. And then need to figure out how to set the quantity.
Need a similar method for remove from list.